Implemented role based auth guards to supermarket's endpoints

+++ src/controller/supermarket.rs
@@ -1,7 +1,10 @@
+use crate::guard::auth_guard::{AuthGuard, Role};
+use crate::library::error_response::error_response;
 use crate::model::supermarket::{CreateSupermarketDto, UpdateSupermarketDto};
 use crate::service::supermarket::SupermarketService;
 use crate::{library::heymart_result::Result, model::supermarket::Supermarket};
-use rocket::{get, post, State};
+use rocket::http::Status;
+use rocket::{get, post, put, State};
 use rocket::{response::status::Created, serde::json::Json};
 use sqlx::PgPool;
 
@@ -23,9 +26,16 @@ pub async fn find_by_id(db_pool: &State<PgPool>, id: i64) -> Result<Json<Superma
 
 #[post("/", format = "json", data = "<supermarket>")]
 pub async fn create(
+    auth_ctx: AuthGuard,
     db_pool: &State<PgPool>,
     supermarket: Json<CreateSupermarketDto>,
 ) -> Result<Created<Json<Supermarket>>> {
+    if auth_ctx.user.role != Role::ADMIN {
+        return Err(error_response(
+            Status::Forbidden,
+            String::from("You are not authorized to perform this action."),
+        ));
+    }
     return match SupermarketService::create(db_pool.inner().clone(), supermarket.into_inner()).await
     {
         Ok(supermarket) => Ok(Created::new("/").body(Json(supermarket))),
@@ -33,12 +43,19 @@ pub async fn create(
     };
 }
 
-#[post("/<id>", format = "json", data = "<supermarket>")]
+#[put("/<id>", format = "json", data = "<supermarket>")]
 pub async fn update(
+    auth_ctx: AuthGuard,
     db_pool: &State<PgPool>,
     id: i64,
     supermarket: Json<UpdateSupermarketDto>,
 ) -> Result<Json<Supermarket>> {
+    if auth_ctx.user.role != Role::ADMIN {
+        return Err(error_response(
+            Status::Forbidden,
+            String::from("You are not authorized to perform this action."),
+        ));
+    }
     return match SupermarketService::update(db_pool.inner().clone(), id, supermarket.into_inner())
         .await
     {
@@ -48,7 +65,13 @@ pub async fn update(
 }
 
 #[delete("/<id>")]
-pub async fn delete(db_pool: &State<PgPool>, id: i64) -> Result<()> {
+pub async fn delete(auth_ctx: AuthGuard, db_pool: &State<PgPool>, id: i64) -> Result<()> {
+    if auth_ctx.user.role != Role::ADMIN {
+        return Err(error_response(
+            Status::Forbidden,
+            String::from("You are not authorized to perform this action."),
+        ));
+    }
     return match SupermarketService::delete(db_pool.inner().clone(), id).await {
         Ok(_) => Ok(()),
         Err(e) => Err(e),
